refactor: improve Custom Elements Manifest outputs
