+++ package.json
@@ -111,9 +111,12 @@
     "@types/react-dom": "18.3.1",
     "autoprefixer": "10.4.20",
     "babel-jest": "29.7.0",
+    "cem-plugin-custom-jsdoc-tags": "1.1.2",
     "cem-plugin-expanded-types": "1.3.3",
+    "cem-plugin-jsdoc-example": "0.0.9",
     "chromatic": "11.12.5",
     "core-js": "3.38.1",
+    "custom-element-jsx-integration": "1.5.3",
     "custom-element-vs-code-integration": "1.4.1",
     "decompress": "4.2.1",
     "eslint": "9.12.0",
@@ -170,10 +173,11 @@
     "vue": "3.5.12"
   },
   "overrides": {
-    "eslint": "9.12.0",
+    "@custom-elements-manifest/analyzer": "$@custom-elements-manifest/analyzer",
     "@jscutlery/semver": {
       "@nx/devkit": "$@nx/devkit"
-    }
+    },
+    "eslint": "9.12.0"
   },
   "volta": {
     "node": "22.11.0",
